backend/requirements.txt
fastapi
uvicorn
python-socketio[asgi]
sqlalchemy
psycopg2-binary
